feat: X-Sentry-Rate-Limits for Events (#449)
The SentryHttpTransport can handle X-Sentry-Rate-Limits only for Events and 429 Retry-After limits for both Events and Envelopes. X-Sentry-Rate-Limits for Envelopes and Events that are queued and sent in sendAllStoredEvents is not supported yet. This is going to be implemented in another PR.
The logic for storing rate limits of both X-Sentry-Rate-Limits and 429 Retry-After is located in SentryDefaultRateLimits. Almost all of the parsing of the HTTP headers was moved to SentryRateLimitParser and SentryRetryAfterHeaderParser.