ref: Remove sendAllEvents from SentryTransport (#465)
Move logic when cached events should be sent into SentryHttpTransport. The sendAllStoredEvents
is removed from SentryTransport. SentryHttpTransport now sends all cached events and envelopes
on init. Furthermore sendAllStoredEvents doesn't only send events, but also envelopes. Therefore
sendAllStoredEvents is renamed to sendCachedEventsAndEnvelopes. The naming in SentryFileManager
reflects that also envelopes are saved in the events path to avoid misunderstandings.
SentryHttpTransport.sendCachedEventsAndEnvelopes currently checks if a RateLimit is active for
events. This causes that sessions are dropped when a RateLimit for events is active. A todo is
added for this problem and it is going to be fixed in another PR.