feat: RateLimiting for cached events and envelopes (#480)
Check if a rate limit is active for cached events and envelopes. The
sending of cached events and envelopes is split up. We don't deserialize
events because we use anyways always the rate limit category error. For
cached envelopes the same logic is applied as for sending envelopes in
the first place.