fix: Overwriting stack trace for crashes (#605)
SentryClient overwrote the stack trace (debug meta and threads) when sending an event that already has debugMeta and threads. This caused stack traces of crashes logged by SentryCrash being overwritten. Now SentryClient keeps 
the threads and debug meta when sending an event and still attaches the stack trace for events that don't have those 
set when attachStacktrace is enabled or that contain an error or exception.