fix: Crash when converting Recrash Report (#627)
SentryCrash sometimes writes recrash reports to the disk that contain an NSString for an element inside crash -> 
threads instead of an NSDictionary. This caused a crash in SentryCrashReportConverter, that is fixed now by discarding
this object in the threads array. Furthermore, the conversion of the report to an event is now surrounded by a try-catch. 
So if a report can't be converted it is just discarded.