fix: Crash in SentryEnvelope.initWithEvent
The serialization to JSON in SentryEnvelope.initWithEvent can lead to a crash when the
serialization of the event returns a non-valid JSON object. This is now avoided by calling
SentrySerialization and checking for an error. If an error in the JSON serialization
occurs an event with a message that the event can't be converted is sent to Sentry.