feat: Use envelopes for sending events
Send events through envelopes to Sentry. SentryHttpTransport maps events to envelopes.
SentryFileManager doesn't store events anymore and deletes the events folder on init so
cached events don't stick around when upgrading from 5.x.x. The handling of trying to
send an event that can't be serialized is done in #643 and therefore we can disable
testSendEventWithFaultyNSUrlRequest for now, but shouldn't release 6.0.0 without this
fix.