fix: Crash in SentryEnvelope.initWithEvent (#643)
The serialization to JSON in SentryEnvelope.initWithEvent can lead to a crash when the
serialization of the event returns a non-valid JSON object. This is now avoided by calling
SentrySerialization and checking for an error. If an error in the JSON serialization
occurs sdk and context are set to nil and the serialization is tried again. If the second serialization succeeds a breadcrumb is to notify the user that the context or sdk could not be serialized. If the second serialization fails a new event with very basic info is created and sent to Sentry to avoid losing the event.