fix: Sending envelopes multiple times (#687)
Sending all cached envelopes in SentryHttpTransport had a concurrency issue. For each envelope to send SentryHttpTransport iterates over all cached envelopes and queues a request. The envelopes get removed from the cache once a request finishes with success. If multiple envelopes get sent in a short time interval SentryHttpTransport iterates over cached envelopes for which already a request is queued. This can lead to multiple requests for the same envelope and other unpredictable behavior as the session stats generated by SentrySessionGeneratorTests being random on Sentry. This issue is fixed now by sequentially sending cached envelopes. When one envelope is sent successfully the next envelope in the cache is queued for sending. If SentryHttpTransport is already sending envelopes and it receives a new envelope the envelope is just stored in the cache and sending all cached envelopes is not triggered again because it is already running. The new cached envelope is going to be picked up later.