feat: Remove SDK frames when attaching stacktrace
Remove SDK frames when attaching the stacktrace for non hard crashes. This doesn't
work though when including the SDK via the Swift Package Manager, because then the SDK
is in the same package as the containing application. We don't want to risk removing too many
frames and therefore we don't remove any frames.