feat: Use error domain and code for event message
When capturing an NSError combine the error message of the error.domain and error.code.
The error.localizedDescription is still set in the context of the event as user info. Issues with the
message coming from the domain and code might not be grouped to existing issues which messages
stem from the error.localizedDescription although there origin is the same. For new issues this
is not going to have a big impact on grouping as it is based upon stack trace, exception, and
message.