ci: Release with Github Actions (#794)
Transfer the release from Travis to Github Actions. To achieve this several workflows are merged together into
buildandtest.yml because the release job should only run if the units tests are green, the samples compile, and Cocoapods and Carthage are validated. This would have been hard to achieve across several workflows.