feat: Bring back SentryOptions.enabled
With 6.0.0 we removed SentryOptions.enabled, but some users used this flag for use cases
of consent management and user privacy to comply with GDPR and related privacy legislation.
This PR brings the enabled flag back with slightly changed functionality. Previously setting
an empty or wrong DSN also set the enabled to false. This side effect is removed now. Setting
the DSN has no impact on the enabled flag. If the DSN is nil or empty or the enabled is set
to false, the client doesn't send anything to Sentry.

Fixes GH-809