perf: Async synching of scope on to SentryCrash
Synchronize the scope to SentryCrash on a background thread instead of the calling thread,
because this can block the calling thread which can be the main thread for a few ms. We
accept the tradeoff that in case of a crash the scope might not be 100% up to date over
blocking the main thread.