feat: Add storeEnvelope on SentryClient (#836)
Hybrid SDKs as react-native need a way to synchronously store an envelope to disk. Previously, this
was possible by exposing SentryFileManager on the SentryClient, but this breaks users of
react-native that use use_frameworks!,
see https://github.com/getsentry/sentry-react-native/issues/1171. A way to fix this would be to expose
SentryFileManager and SentryCurrentDateProvider as proposed in
https://github.com/getsentry/sentry-cocoa/pull/835. This has the downside of making classes public,
that shouldn't be public. A better workaround is to expose storeEnvelope on the SentryClient.