fix: Serialization of SentryScope
The SentryScope contains several properties that are dictionaries or arrays. The serialize method
returned a dictionary with references. So when for example a tag is added to the scope it is also
added to the serialized dictionary of the scope. This can lead to crashes when iterating over the
tags of the serialized dictionary while another thread adds a tag to the scope. This is fixed now
by adding a copy of these properties to the serialization dictionary instead of adding them by
reference.