fix: Remove SentryScope.isEqual and hash
Adding these implementations was a mistake and they should not be there. We only use the
isEqual in testing code and therefore we can move it there. Strictly speaking, this is a breaking
change, but we can't think of a use case where someone would use them. We can easily add
them again.