ref: Use instance variable in Hub to access Client
We use [self getClient] to access the client in the hub. This is not needed and can be replaced by
accessing the instance variable directly via _client. This makes the code easier to read and also
slightly faster as we don't have to go through Objective-C method dispatch when accessing the
client via a property.