fix: Capture envelope updates session state
The SentryHub now increases the error count if an envelope is passed containing an event with
event.level error or higher. Ideally, we would check the mechanism and/or exception list, like the
Java and Python SDK do this, but this would require full deserialization of the event.

Fixes GH-861