fix: Sessions for Hybrid SDKs
The Cocoa SDK didn't start a session after a hybrid app finished its initialization. That's because, in hybrid SDKs, we might initialize the Cocoa SDK after the hybrid engine is ready. The SessionTracker then misses the didBecomeActive notification from the OS and fails to start a session. To solve this, hybrid SDKs can post an internal notification with the name SentryHybridSdkDidBecomeActive. The SessionTracker starts a session if it receives such a notification.