test: Fix ThreadInspectorTests (#927)
The SentryThreadInspectorTests.testThreadNameIsNull was sometimes failing in CI. The
thread name was set to some weird characters as ȕ`\r\u{01}. This was because we used
empty array to fill the buffer for the thread name and therefore accessed memory, which
we shouldn't. Now we use an array with all zeros as big as the buffer.