fix: Crash in Session Init with JSON
When the JSON contains a started that can't be converted to a NSDate the SDK sets
the _started to nil, which leads to a crash later as _started is non null. This is fixed now
by not setting the _started if the SDK can't convert the started to a NSDate.

Fixes GH-937