feat: Capture async stack traces
libdispatch on Apple has its own introspection support. However in order to use that, the user has to actually load the introspection capable version of the library, which is the case when doing debugging in XCode, but is not the case for normal runtime.

See https://github.com/apple/swift-corelibs-libdispatch/blob/main/dispatch/introspection.h

Instead, we rely on https://github.com/facebook/fishhook to replace the async libdispatch functions with our own wrappers, similar to how https://github.com/tiantianbobo/XBAsyncStackTrace works.