feat: Capture async stack traces (#998)
libdispatch on Apple has its own introspection support. However in order to use that, the user has to actually load the introspection capable version of the library, which is the case when doing debugging in XCode, but is not the case for normal runtime.

See https://github.com/apple/swift-corelibs-libdispatch/blob/main/dispatch/introspection.h

Instead, we rely on https://github.com/facebook/fishhook to replace the async libdispatch functions with our own wrappers, similar to how https://github.com/tiantianbobo/XBAsyncStackTrace works.

These wrappers will then eagerly capture a stacktrace in the caller, which is then moved into a wrapper block that is submitted to the real libdispatch. Inside that block, the async stacktrace is saved in a thread-local which is being used when chaining nested async calls, or when actually capturing a crash stacktrace.

Those chained stacktraces are then being walked when iterating the stacktrace.

Co-authored-by: Philipp Hofmann <philipp.hofmann@sentry.io>
Co-authored-by: Bruno Garcia <bruno@brunogarcia.com>