perf: Scope sync to SentryCrash (#1193)
On each change to the scope, the SDK serialized the whole scope and synced it to SentryCrashed. Due to this big overhead, the synchronization was done on a background thread. This is fixed now, but having a fine-grained sync. Whenever a property in the scope changes, the SDK only synchronizes this property to SentryCrash. The SDK serializes the property to JSON, which is stored in C. When a crash happens, the SentryCrashReport picks up the JSON of all properties and adds it to the crash report. On app start, the SentryCrashReportConverter merges the userInfo with the sentry_sdk_scope to be backward compatible with older crash reports and to minimize the required changes.

Fixes GH-991