fix: AppStartMeasurement didFinishLaunching is nil (#1228)
In some circumstances, it can happen that didFinishLaunching in SentryAppStartTracker
is not called because adding the observer to UIApplicationDidFinishLaunchingNotification
is happening after the specified notification. This results in didFinishLaunchingTimestamp
being nil. When the SentryTracer then tries to add the app start spans and
measurements, the spans are invalid. The SDK sends the transaction to Sentry, but it
never appears. This is fixed now.

Fixes GH-1225