fix: Swizzle only inApp ViewControllers (#1242)
SentryUIViewControllerSwizziling was checking the image name of the app delegate class.
For SwiftUI apps without a specified app delegate, the app delegate is part of the SwiftUI
framework. Therefore SentryUIViewControllerSwizziling also swizzled ViewControllers of
SwiftUI which is not desired. This is fixed now by making use of the existing
SentryInAppLogic.

Fixes GH-1229