fix: Crash when swizzling Nib UIViewController (#1277)
The SDK always swizzled the loadView method of an implementation of UIViewController.
When using an UIViewController with a nib, UIKit only loads the nib if the loadView method
is not overridden. Our swizzling confused the UIKit, and therefore, it didn't load the nib
leading to an empty view and then to a crash. This is fixed now by not swizzling
UIViewControllers with a nib file if they don't implement loadView.