fix: Sentry HTTP Trace Header Breaking Requests (#1295)
Removed SentryURLProtocol that was responsible for adding trace header in every request. The startLoading
used NSURLSessionConfiguration.defaultSessionConfiguration, which didn't include custom protocol classes.
This is fixed now by swizzling 
NSURLSessionConfiguration.HTTPAdditionalHeaders. Whenever this getter is
called, the SDK adds the trace header if there is a span on the scope.

Fixes GH-1288

Co-authored-by: Sentry Github Bot <bot+github-bot@sentry.io>
Co-authored-by: Philipp Hofmann <philipp.hofmann@sentry.io>