fix: Swizzling crash on iOS 13
iOS 13 doesn't have a method for HTTPAdditionalHeaders. Instead, it only
has a property. Therefore, the swizzling crashes on iOS 13. This is fixed now
by checking if NSURLSessionConfiguration has a method HTTPAdditionalHeaders.
The downside is that we don't add a trace header on iOS 13 anymore. This
is better than crashing though.

Fixes GH-1296