fix: HTTP instrumentation KVO crash
The HTTP instrumentation in SentryNetworkTracker leads to EXC_BAD_ACCESS.
This problem could be due to not correctly removing the KVO observer in
SentryNetworkTracker. Other open-source projects had similar bugs due to
using KVO, which they fixed using swizzling instead. As swizzling doesn't work
for our current solution and removing KVO would be a lot of effort, this PR is the
best shot effort for fixing the crashes by removing the observer issue.

Fixes GH-1328