fix: Crash When Observing Span Finished
The tracer observes when one of its child spans finishes when waitForChildren
is enabled. A prior fix for memory leaks GH-1352 led to crashes when watching
the child spans with KVO. This problem is fixed now by replacing KVO with a
simple method call from span to tracer when the span finishes.

Fixes GH-1358