test: Fix SentryStacktraceBuilderTests
testAsyncStacktraces should be testing if stack traces of different threads
are stitched together, which it didn't do, because we were only building
the stack trace on the main thread. This is fixed now by building the
stack trace across different threads. Furthermore, the tests now use
its own queue to avoid side effects when testing.