fix: Crash for Call Should be on Main Thread (#1371)
The swizzling of subclasses of UIViewController stored
references to classes on a background thread, which led
to calling the initialize method. Some UIViewControllers
implementing assume that they are on the main thread
when executing initialize. Calling initialize from a
background thread can lead to crashes. This is fixed now
by avoiding the call to initialize on a background thread and
swizzling on the main thread.

Fixes GH-1366