test: Fix flaky UrlSessionDelegate test (#1397)
The test for the URLSessionDelegate was making a web request to Sentry.
When the DSN of the SDK is rate limited the test failed. Instead, now we test
if the delegate is correctly passed down to the SentryQueueableRequestManager
in the SentryTransportFactory and we make a normal request to github.com.