fix: Crash in Span when Tracer nil (#1416)
The SentrySpan has a weak reference to SentryTracer to avoid a strong
reference cycle. Therefore, ARC can deallocate the SentryTracer if no one
keeps a strong reference to the SentryTracer. When calling startChild on
SentrySpan and the SentryTracer are nil, the code currently crashes. This
is fixed now by adding a SentryNoOpSpan.

Fixes GH-1367