fix: SentryOptions initWithDict type errors (#1443)
When initializing the SentryOptions with a dictionary, you can pass in a different type
for a property than expected. The SentryOptions had some checks validating if the
types from the dictionary match the actual type of the property. Anyways, the logic
missed a few edge cases as NSNull or didn't check if callbacks are blocks. This is
fixed now by checking if all properties are not NSNull and if callbacks are of type block.

Fixes GH-1441