feat: Send enabled SDK integrations
Send the list of enabled integrations in the event.sdk payload
to know which integrations are enabled. Cause we always install
integrations no matter if they get enabled or not, this PR adds
isEnabled to the integration protocol. With this integrations
can communicate if they are enabled or not.

Fixes GH-1646