test: Fix flaky SentryClientTests (#1709)
The SentryClientTests used SentryThreadInspector for asserting that SentryClient
appropriately calls the SentryThreadInspector, and it also validated the thread's
stacktraces. The assertion also compared the number of threads, which can easily
change between the call to SentryClient and the assertion later on. As
SentryThreadInspectorTests test the functionality of SentryThreadInspector, we can
instead use a test implementation in SentryClientTests to make the tests deterministic.

Fixes GH-1706