fix: Setting invalid sample range
The hub discarded transactions when the sample decision was not yes
instead of yes. The TracesSampler returned no when the SDK should
sample the transaction. While the traces sample rate worked when
setting a valid range, the TracesSampler returned the sampling decision
no when the sample range was invalid, leading to sampling all
transactions when the sample range was invalid. This is fixed now.