feat: Collect queue label information for profiles (#1828)
We previously avoided collecting queue names because the dispatch_qaddr field (see diff) would point to a queue that was no longer valid by the time we read the field, causing a crash. I found a note in some Apple code that explains what was happening here: https://github.com/WebKit/WebKit/blob/09225dc168d445890bb0e2a5fa8bc19aef8556f2/Source/WebCore/page/cocoa/ResourceUsageThreadCocoa.mm#L119

The queue had been deallocated by the time we were reading the field. In order to avoid this race condition, in this PR, we suspend the thread before trying to read the field from the queue. Suspending the thread should prevent the dispatch queue from deallocated, assuming the deallocation happens on the thread that the queue is currently assigned to. I am not yet 100% confident in this behavior but we will have to try this in production (for opt-in profiling beta customers only, this doesn't impact anything else) and see if this works.

Each sample has a queue_address key that can be used to look up the corresponding queue label from the top level queue_metadata object.