Update SentryCrashDefaultMachineContextWrapper.m (#1928)
We were using `+initialize` method from `SentryCrashDefaultMachineContextWrapper` to save the main thread id. `+initialize` is called with the first message sent to the class, and this can occur in any thread.

'+load' is called when the library is load and this always occurs on the main thread, therefore a much safer place to save main thread id.