fix: Respect existing baggage header instead of overwriting it (#1995)
SentryNetworkTracker adds a baggage header with some sentry keys and values, but it does so by completely overwriting any possibly existing baggage header. And when we remove the baggage header when SentrySDK.currentHub.scope.span is nil, we also do that by removing the entire baggage header instead of only the sentry-trace_id key that we actually need to remove.

Now we respect the existing baggage header, if it's there, by modifying it.