ref: remove integrations shouldBeDisabled duplication (#2032)
There is a lot of duplication in all the integration, in the shouldBeDisabled code. They all check a bunch of options, and the logging logic is all the same. They all need to remove the integration from the installed options as well. Now, everything is simplified with a base class, and we only have to give an array of integration options.