fix: standardize on C++14 (#2068)
We should standardize on the version used to build the framework, as that'll help avoid weird issues going forward if any experimentation is done in a sample app or test target. We can always change this if there is a good reason to in a particular target in the future.