feat: Allow error event value (description) to be customised (#2120)
Currently, errors are hard coded to use "Code: xxx" using the error's code which makes for a very
unhelpful display in the Sentry error list. To alleviate this, check for the presence of a custom
description in the error's userInfo dictionary using the NSDebugDescriptionErrorKey. If no custom
description is set, the old behavior is retained by using "Code: xxx".

Co-authored-by: Matt Smollinger <msmollinger@community.com>
Co-authored-by: msmollin-community <62894137+msmollin-community@users.noreply.github.com>