ci: Remove caching for UI tests (#2210)
The UI tests were using caches that stopped the UI tests from
recompiling when there were changes in the library code.
Instead, we need to recompile the test sample apps with their
UI tests when there are changes in the library code to be able
to catch bugs. This is fixed now by removing the cache.