fix: Add app start to first finished transaction
The SDK added the app start measurement and spans to the first
finished auto-generated UIViewController transaction. The first finished
UIViewController transaction is not necessarily the transaction of the
first loaded screen. As auto-generated transactions for
UIViewControllers wait for all children to finish, it may kick off, for
example, a web request that lasts for a couple of seconds in the
background. In the meantime, the user navigates to another screen,
which finishes loading before the web request. The SDK added the
app start data to the wrong transaction in such an edge case. This is
fixed now by adding the app start data to the first started transaction.

Fixes GH-2248