feat: MetricKit TestVersion 1 (#2476)
This PR adds a first iteration of sending MetricKitData of MXCrashDiagnostic, MXCPUExceptionDiagnostic, and MXDiskWriteExceptionDiagnostic to Sentry. This PRs aim is to have the code running in the iOS-Swift test app via TestFlight to get some real MetricKit data of MXCPUExceptionDiagnostic, and MXDiskWriteExceptionDiagnostic.

MXCrashDiagnostic exists solely for validating the stacktrace symbolication.