fix: Remove delay for deleting old envelopes
The SentryFileManager deleted old envelopes after 10 seconds,
which resulted in accessing double-freed pointers. We delayed the
deletion for 10 seconds to avoid additional operations when starting
the SDK. As the overhead of deleting the old envelope items should
be minimal, we remove the delay because this is simpler than not
running the delayed deletion for tests and reduces complexity.