test: Fix flaky testFaultyReportIsNotSentAndDeleted
The test was validating the wrong invocation on the test client.
Instead of captureEventWithScopeInvocations it should have been
captureCrashEventInvocations. Anyways, we can improve the test
by not validating no events being passed down to the client, but instead
use the completion callback of the SentryCrashInstallationReporter to
get rid of the delayNonBlocking.