fix: Cleanup AppHangTracking properly when closing SDK
When closing the SDK directly after starting it, it could happen that the ANRTracker
didn't start its watchdog thread yet. Canceling the ANRTrackers thread required the
instance of the watchdog thread and didn't work correctly. This is fixed now by using
simple state management in the ANRTracker. While this is an edge case, it can help
with flaky tests, as the test logs showed signs of the ANRTracker running in the
background.