fix: Serialization of nullable booleans (#2706)
The protocol contains a couple of nullable BOOLs that are of type
NSNumber. Setting these properties to something different than 0
or 1 leads to a serialized NSNumber, which violates the protocol.
With this refactoring 0 value always means false, and any nonzero
value is interpreted as true when serializing such nullable BOOLs.