ref: Add delegate for system event crumbs (#2709)
Add a delegate for SentrySystemEventBreadcrumbs to be independent of SentrySDK to
make the validation of crumbs easier in SentrySystemEventBreadcrumbsTests, and
only one test in SentryAutoBreadcrumbTrackingIntegrationTests has to use the global
state of the SentrySDK for testing to reduce flakiness. Furthermore, use test classes
in SentryAutoBreadcrumbTrackingIntegrationTests when calling install to reduce
side effects when testing.