convert remaining profiling sources from pure C++ to objective C++
This will allow us to start using Cocoa types in the sources. I ran into this trying to convert getDurationNs to return an NSNumber, so it can return nil if the calculation would product a negative duration (which is not allowed since it currently returns a uint64_t). But including Foundation breaks the C++ compiler because SentryTime.h is included in SentryBacktrace.cpp.