fix: profiler GPU timestamp assumptions
originally assumed that the 'previous' and 'current' times in the CADisplayLink callback represented the start and end time of a rendered GPU frame. They are actually just the times in between calls to render the frame; the duration of the actual render is always the same. What causes longer times between callbacks are the main thread stalls, which is really what we want to record. So, always record the 'current' time for the timestamp, instead of 'previous' which we did thinking that was the actual start time of the frame; still compute duration the same way.

Consolidate how GPU info are recorded and sliced in the frames tracker and profiler, respectively.