test: FileManager remove perf test (#2869)
CI can't compare the performance test with measure against anything, as Xcode
needs baseline files for the same computer. In CI, the computer specifications might
change frequently, so it's a lot of effort to create those baseline files. This test only
slows down CI, and therefore, we can remove it. For more context see
https://github.com/getsentry/sentry-cocoa/issues/1482.