ci: Remove schedule of unit tests (#3030)
We hypothesized that tests take longer and fail more often when the US is
online. We didn't find strong evidence for this claim. There could be a slight
correlation between the US being online and the tests failing more often.
Even with a weak correlation, there isn't much we can do except improve the
test suite. Therefore, keeping the tests running on a schedule doesn't make
sense.