ci: always run setup_ci
For some reason if these jobs are both run on ci sequentially, the second one fails in the match phase. I think setup_ci lets the previous run's keychain get used for the subsequent run's, and then something fails. In any case, these are the only jobs that don't call this with force: true.