Fix: Remove sync call to main thread from SentryUIDeviceWrapper (#3295)
SentryUIDeviceWrapper previously had a synchronized call to the main thread when the class was initialized in a background thread, potentially leading to deadlocks with SentryDependencyContainer.

To address this, the synchronized call has been replaced with an asynchronous one. Additionally, a new start function has been introduced and is invoked from SentrySDK.start to ensure initialization as early as possible.