fix: Crash in SentryTracer when cancelling timer
The SentryTracer starts the timer for the deadline on the main thread
and invalidates the timer, not necessarily on the main thread, but
invalidate has to be called on the same thread on which the timer was
started. This is fixed by starting and invalidating the NSTimer on the
main thread.

Fixes GH-3320