fix: Multiple issues in Reachability
The tests can launch multiple threads of the ANR tracker, which call
dealloc of SentryReachability simultaneously, leading to a segfault
while iterating over the observers in SentryReachability. This is fixed
now by using some synchronization around the observer's access.
Furthermore, the bookkeeping of the observers is implemented now with
weak references using an NSHashTable. Instead of relying on dealloc to
be called for removing the observers, each observer must now call
removeObserver after subscribing. If an observer doesn't call
removeObserver no memory is leaked cause the SentryReachability uses
weak references. Previously, the list of observers used a dictionary
with the description method of Objective-C as the key, which is
suboptimal cause the description could change between adding and
removing the observer, leading to memory leaks when the observer's
description changed between invocations.