Fix: SDK detects OutOfMemory after device reboot (#3352)
If the host app has been terminated due to device reboot, SentrySDK assumes it as a WatchdogTermination crash. In such cases, mostly, the boot time will be mentioned as "a few seconds before this event". 

![Screenshot 2023-10-19 at 09 52 27](https://github.com/getsentry/sentry-cocoa/assets/1960884/c66f778a-cb4e-4687-991b-646de004b194)

The idea behind this fix is that if the system boot time has been changed since the previous app launch, most likely the app has been terminated due to the reboot and the SentrySDK should not consider it as a WatchdogTermination crash.

Co-authored-by: Dhiogo Brustolin <dhiogo.brustolin@sentry.io>