test: Assert clearTestState on main thread
Add an assert for checking clearTestState executed on the main thread. I
suspected one of our tests was flaky because we weren't running
clearTestState on the main thread, which was a false assumption. I added
this assert to avoid any future problems caused by running
clearTestState on a background thread.