ref: Add UIEventTracker Mode (#3406)
Carrier transactions are going to replace UI event transactions in the
future. When users enable carrier transactions, UI event transactions
must be disabled because both would interfere. Therefore, prepare the
SentryUIEventTracker to be able to handle two different modes—one for
the existing functionality of UI event transactions and one for adding
spans to carrier transactions.