fix: Crash in SentryFramesTracker
The SentryFramesTracker registers a callback to the CADisplayLink. While the SentryFramesTrackingIntegration keeps a reference to the SentryFramesTracker and calls SentryFramesTracker.stop on uninstall, there could be situations where the SentryFramesTracker is deallocated, but not calling invalidate on the CADisplayLink. Then the CADisplayLink still calls  and crashes with EXC_BAD_ACCESS cause the frames tracker is deallocated. To fix this problem SentryFramesTracker.dealloc now calls invalidate on the CADisplayLink.