ref: Ignore getFramesDelay for ThreadSanitizer
The ThreadSanitizer now ignores SentryFramesTracker:getFramesDelay.  We
accept the data race of the two properties _currentFrameRate and
previousFrameSystemTimestamp, that are updated on the main thread in the
displayLinkCallback. This method only reads these properties. In most
scenarios, this method will be called on the main thread, for which no
synchronization is needed. When calling this function from a background
thread, the frames delay statistics don't need to be that accurate
because background spans contribute less to delayed frames. We prefer
having not 100% correct frames delay numbers for background spans
instead of adding the overhead of synchronization.