fix: Make SentryScope.useSpan non-blocking (#3568)
Instead of calling the callback of useSpan within the spanLock, get a 
reference to the span inside the critical sector and pass this to the 
callback after the critical sector closes. This way, when the callback 
is a blocking call, useSpan doesn't block other invocations of itself 
and prevents potential deadlocks.