fix: Rare battery breadcrumbs crash (#3582)
We see rare crashes in getBatteryStatus of SentrySystemEventBreadcrumbs.
The object of an NSNotification can be nil, and we didn't check for that
edge case. This PR fixes that problem and uses dictionaryWithCapacity to
slightly reduce the memory footprint for getBatteryStatus. This change
is not guaranteed to fix these rare crashes, but it won't do any harm.