fix: FileManager log info when path doesn't exist (#3594)
Logging an error when getting all files for a folder that doesn't exist
confuses users; see GH-3577. Instead, log an info message that the
SentryFileManager tried getting a list of files for a folder that
doesn't exist. We chose info because it's still an edge case that
shouldn't occur often, but it doesn't break any functionality. If a path
was deleted manually by the user that the SDK requires, the SDK will
create the path on next launch. We don't want to add checks if all the
necessary paths exist every time the SDK stores an envelope cause this
adds some overhead. If users manually remove the required paths, we
accept that they have to wait until the next time the SDK launches.