fix: Move headers out of Extern C (#3690)
If a project uses a C++ compiler, it will not compile because of some headers inside a "extern C" block referencing C++ code.

This appears to be an inheritance from KSCrash that I don't fully understand the reasoning behind. Moving these headers out of the extern block hasn't caused any crashes in our default pipeline, so I believe we're fine with it.